+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Remove Advisory component
+
 ## 8.8.2
 
 * Fix a bug where headings were no longer rendering inside example boxes ([#374](https://github.com/alphagov/govspeak/pull/374))